pyproject.toml
[project]
name = "flashcards"
version = "0.1.0"
description = "Flash Card Game for Medical Terminology"
requires-python = ">=3.10"
dependencies = [
    "python-bidi>=0.4.2",
]

[project.scripts]
flashcards = "flashcards.app:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
flashcards = ["*.md"]
